# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/teamlead/dashboard": {
      "get": {
        "operationId": "TeamleadController_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/board": {
      "get": {
        "operationId": "TeamleadController_board",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "§10.3 Mitarbeitenden-Board for the day (assigned bundles per employee)",
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/capacity": {
      "get": {
        "operationId": "TeamleadController_capacity",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapacityDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "§10.1 Day capacity tile (net / planned / free / utilisation)",
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/kpis": {
      "get": {
        "operationId": "TeamleadController_kpis",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KpiDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "§10.1 Day ZST KPIs (computed from ZstRecord + case statuses)",
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/events": {
      "get": {
        "operationId": "TeamleadController_events",
        "parameters": [
          {
            "name": "actorType",
            "required": false,
            "in": "query",
            "description": "Filter by ActorType (system|employee|teamlead|admin)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "description": "Filter by entityId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "description": "Comma-separated WorkflowEventType allowlist, e.g. \"assignment.overridden,case.prioritized,case.parked,case.ready\". When set, only events whose eventType is in the list are returned.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size 1..200",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEventDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "§7.2/§16.2 audit feed (workflow events, newest first)",
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/cases": {
      "get": {
        "operationId": "TeamleadController_pool",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by CaseStatus",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "section",
            "required": false,
            "in": "query",
            "description": "Filter by section code",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Volltext: WE-Beleg-Nr / Lagerplatz-Code / Lieferschein-Nr (contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shopNo",
            "required": false,
            "in": "query",
            "description": "Filter: Shop (primärer Shop, contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchNo",
            "required": false,
            "in": "query",
            "description": "Filter: Filiale (contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bereich",
            "required": false,
            "in": "query",
            "description": "Filter: fester Bereich (Hängebahn|Palette|Regal)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assigned",
            "required": false,
            "in": "query",
            "description": "Filter: zugeteilt ja/nein",
            "schema": {
              "type": "string",
              "enum": [
                "yes",
                "no"
              ]
            }
          },
          {
            "name": "labels",
            "required": false,
            "in": "query",
            "description": "Filter: Etiketten nötig ja/nein",
            "schema": {
              "type": "string",
              "enum": [
                "yes",
                "no"
              ]
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "Lebenszyklus-Scope (server-seitig): aktiv | abgeschlossen | archiv (completed+zst_done) | topf (Aufmerksamkeit/blocked/needs_review) | alle",
            "schema": {
              "type": "string",
              "enum": [
                "aktiv",
                "abgeschlossen",
                "archiv",
                "topf",
                "alle"
              ]
            }
          },
          {
            "name": "bookingFrom",
            "required": false,
            "in": "query",
            "description": "Buchungsdatum ab (YYYY-MM-DD, inklusive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingTo",
            "required": false,
            "in": "query",
            "description": "Buchungsdatum bis (YYYY-MM-DD, inklusive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Server-side sort column",
            "schema": {
              "type": "string",
              "enum": [
                "weBelegNo",
                "bookingDate",
                "totalQuantity",
                "effortPoints",
                "status",
                "section",
                "branchNo",
                "primaryShopNo",
                "completedAt"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the operational pool (filter + paginate)",
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/cases/lookup": {
      "get": {
        "operationId": "TeamleadController_lookupCase",
        "parameters": [
          {
            "name": "weBelegNo",
            "required": true,
            "in": "query",
            "description": "WE-Belegnummer (exakt, case-insensitive)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseLookupResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "B1 WE-Nr-Zuweisung: look a Beleg up by WE-Belegnummer with an assignability verdict",
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/cases/search": {
      "get": {
        "operationId": "TeamleadController_searchCases",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Volltext: WE-Beleg-Nr / Lieferschein-Nr / Lagerplatz-Code / Shop / Filiale (contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bereich",
            "required": false,
            "in": "query",
            "description": "Filter: fester Bereich (Hängebahn|Palette|Regal)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shopNo",
            "required": false,
            "in": "query",
            "description": "Filter: Shop (primärer Shop, contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchNo",
            "required": false,
            "in": "query",
            "description": "Filter: Filiale (contains)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max. Ergebnisse (1-50)",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseSearchResultDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A1/A2/B1: bounded, ranked search + browse over the assignable pool (ready + unassigned) for AssignDialog",
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/cases/{caseId}": {
      "get": {
        "operationId": "TeamleadController_caseDetail",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "§10.4 Belegdetails: one case with positions, boxes, documents, history",
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/cases/{caseId}/prioritize": {
      "post": {
        "operationId": "TeamleadController_prioritize",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrioritizeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "teamlead"
        ]
      }
    },
...
    "/api/teamlead/cases/{caseId}/return-to-bucher": {
      "post": {
        "operationId": "TeamleadController_returnToBucher",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnToBucherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "D1 „Zurück an Bucher\": blockierten Beleg an den Bucher melden (mock Queue)",
        "tags": [
          "teamlead"
        ]
      }
    },
    "/api/teamlead/cases/{caseId}/complete-intake": {
      "post": {
        "operationId": "TeamleadController_completeIntake",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteIntakeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "D1 Freigabe: fehlende Pflichtfelder nachtragen; vollständig → blocked → ready",
        "tags": [
          "teamlead"
        ]
      }
    },
...
    "/api/teamlead/cases/{caseId}/unflag-attention": {
      "post": {
        "operationId": "TeamleadController_unflagAttention",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A7 TL-Topf: Aufmerksamkeitsflag entfernen („aus Topf entlassen\")",
        "tags": [
          "teamlead"
        ]
      }
    },
...
  },
  "info": {
    "title": "Paketlagerdispo Backend API",
    "description": "Digitale Belegverteilung – modular monolith backbone. Status state machine (§7), RBAC (§5/§16.1), audit event log (§7.2), and the case pool/lifecycle (§14.2).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "OIDC access token"
      }
    },
    "schemas": {
...
      "DashboardDto": {
        "type": "object",
        "properties": {
          "countsByStatus": {
            "type": "object",
            "description": "Open case count grouped by status"
          },
          "poolSize": {
            "type": "number"
          },
          "prioOpen": {
            "type": "number"
          },
          "oldestOpenBookingDate": {
            "type": "object",
            "nullable": true,
            "description": "Booking date of oldest open case"
          },
          "endOfShiftOpenCount": {
            "type": "number",
            "description": "Non-terminal Belege assigned to an employee whose shift has already ended (Punkt 6 — offen am Schichtende)."
          }
        },
        "required": [
          "countsByStatus",
          "poolSize",
          "prioOpen",
          "endOfShiftOpenCount"
        ]
      },
...
      "PrioritizeDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Reason for manual teamlead priority"
          }
        }
      },
...
      "ReturnToBucherDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "description": "Hinweis an den Bucher"
          }
        }
      },
...
    }
  }
}